Enhanced watchmode
Allow an external program to be specified to watch particular games
to make use of NH4's wachmode. Add game config options watch_path,
watch_args and watchcommands to support this.  Also add %w/%W for
format tokens the watched player, and %l/%L for the logged-in user
because %n/%N fall back to the watched player if nobody is logged
in, which is not always what we want.  Watch menu now displays N/A
for the terminal size of games which use this, because NH4 watch
mode does not care about player's terminal dimensions.  Other external
watching tools may need this in the future, but for now there is
nothing that requires this.